requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

pymongo==4.6.0
motor==3.3.2

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2

pydantic==2.5.0
pydantic-settings==2.1.0

numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
